# docs/architecture/openapi.yaml
...
openapi: 3.1.0
info:
  title: API Análisis Predictivo — Desarrollos Comerciales del Sur, S.A.
  description: Contrato del prototipo de tesis. Autenticación JWT y control de acceso RBAC.
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: Desarrollo local
paths:
...
  /api/v1/predictions/demand:
    post:
      tags:
      - Predictions
      summary: Inferencia de demanda por producto/ruta/horizonte
      description: Requiere permiso `prediccion:consultar`. Usa el modelo en estado `produccion` (400
        `SIN_MODELO_PRODUCTIVO` si no existe). El intervalo es al 95 %. Con `persistir=true` guarda los
        puntos en `pronosticos_demanda` (idempotente). 400 `HISTORIAL_INSUFICIENTE` si el producto/ruta
        no tiene historial previo a `fecha_base`; 500 `ARTEFACTO_CORRUPTO` si el hash del artefacto no
        coincide (no se sirve predicción).
      operationId: predecir_demanda_api_v1_predictions_demand_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DemandRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemandResponse'
        '400':
          description: '`SIN_MODELO_PRODUCTIVO` o `HISTORIAL_INSUFICIENTE`'
        '500':
          description: '`ARTEFACTO_CORRUPTO`: hash del artefacto inválido; alerta crítica'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/v1/liquidaciones:
    post:
      tags:
      - Liquidaciones
      summary: Crear o actualizar el borrador de la liquidación del día
      description: 'Requiere `liquidaciones:registrar`. Un borrador puede guardarse incompleto y **no**
        afecta al modelo. 400: `VENTA_FECHA_FUTURA`, `PRODUCTO_DUPLICADO_EN_CIERRE`, `PRODUCTO_INACTIVO`,
        `UNIDADES_NO_CUADRAN` (vendido + devuelto + merma excede lo cargado), `CARGA_REQUIERE_JUSTIFICACION`.
        409 `LIQUIDACION_DUPLICADA` si ya está cerrada.'
      operationId: guardar_borrador_api_v1_liquidaciones_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LiquidacionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiquidacionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Liquidaciones
      summary: Historial de liquidaciones
      description: Requiere `liquidaciones:leer` (Administrador y Gerente). Filtros por rango de fechas,
        ruta, vendedor y estado; paginado, más recientes primero.
      operationId: listar_api_v1_liquidaciones_get
      security:
      - HTTPBearer: []
      parameters:
      - name: desde
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Desde
      - name: hasta
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Hasta
      - name: ruta_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Ruta Id
      - name: vendedor_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Vendedor Id
      - name: estado
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/EstadoLiquidacion'
          - type: 'null'
          title: Estado
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListadoLiquidaciones'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/liquidaciones/precarga:
    get:
      tags:
      - Liquidaciones
      summary: 'Precarga de la pantalla: carga despachada y liquidación vigente'
      description: Requiere `liquidaciones:registrar`. Devuelve las presentaciones con la cantidad despachada
        a la ruta ese día (editable, con justificación) y la liquidación vigente de la fecha/ruta si existe.
        Sin carga despachada devuelve una advertencia, no un error.
      operationId: precargar_api_v1_liquidaciones_precarga_get
      security:
      - HTTPBearer: []
      parameters:
      - name: fecha
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Fecha
      - name: ruta_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Ruta Id
      - name: vendedor_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Vendedor Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrecargaResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/liquidaciones/config:
    get:
      tags:
      - Liquidaciones
      summary: Umbral de diferencia de caja
      description: Requiere `liquidaciones:leer`.
      operationId: obtener_config_api_v1_liquidaciones_config_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiquidacionConfig-Output'
      security:
      - HTTPBearer: []
    put:
      tags:
      - Liquidaciones
      summary: Actualiza el umbral de diferencia de caja
      description: Requiere `liquidaciones:corregir` (Administrador). Auditado.
      operationId: actualizar_config_api_v1_liquidaciones_config_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LiquidacionConfig-Input'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiquidacionConfig-Output'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/liquidaciones/{liquidacion_id}:
    get:
      tags:
      - Liquidaciones
      summary: Detalle de una liquidación
      description: Requiere `liquidaciones:leer`. 404 `LIQUIDACION_NO_ENCONTRADA`.
      operationId: obtener_api_v1_liquidaciones__liquidacion_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: liquidacion_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Liquidacion Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiquidacionResponse'
        '404':
          description: '`LIQUIDACION_NO_ENCONTRADA`'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/liquidaciones/{liquidacion_id}/cerrar:
    post:
      tags:
      - Liquidaciones
      summary: Cerrar la liquidación (alimenta el modelo)
      description: 'Requiere `liquidaciones:cerrar`. Valida el cuadre de unidades (400 `UNIDADES_NO_CUADRAN`)
        y de dinero (400 `MONTOS_NO_CUADRAN`); la diferencia de caja no bloquea: se registra y, sobre
        el umbral, genera una alerta. En una sola transacción escribe `ventas_historicas` (origen `liquidacion`),
        rellena `demanda_real`, encola `evaluate_production` y audita. 409 `LIQUIDACION_YA_CERRADA`, `LIQUIDACION_YA_ANULADA`,
        `VENTAS_EXCEL_EXISTENTES`.'
      operationId: cerrar_api_v1_liquidaciones__liquidacion_id__cerrar_post
      security:
      - HTTPBearer: []
      parameters:
      - name: liquidacion_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Liquidacion Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiquidacionResponse'
        '404':
          description: '`LIQUIDACION_NO_ENCONTRADA`'
        '409':
          description: Conflicto de estado de la liquidación
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/liquidaciones/{liquidacion_id}/corregir:
    post:
      tags:
      - Liquidaciones
      summary: Corregir una liquidación cerrada
      description: Requiere `liquidaciones:corregir`. Incrementa `version`, hace UPSERT en `ventas_historicas`,
        recalcula `demanda_real` y vuelve a encolar la evaluación; deja auditoría con el motivo. 409 `LIQUIDACION_NO_CERRADA`.
      operationId: corregir_api_v1_liquidaciones__liquidacion_id__corregir_post
      security:
      - HTTPBearer: []
      parameters:
      - name: liquidacion_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Liquidacion Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorreccionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiquidacionResponse'
        '404':
          description: '`LIQUIDACION_NO_ENCONTRADA`'
        '409':
          description: Conflicto de estado de la liquidación
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/liquidaciones/{liquidacion_id}/anular:
    post:
      tags:
      - Liquidaciones
      summary: Anular una liquidación
      description: Requiere `liquidaciones:corregir`. Motivo obligatorio. Si estaba cerrada revierte su
        efecto en `ventas_historicas` y `demanda_real`. 409 `LIQUIDACION_YA_ANULADA`.
      operationId: anular_api_v1_liquidaciones__liquidacion_id__anular_post
      security:
      - HTTPBearer: []
      parameters:
      - name: liquidacion_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Liquidacion Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnulacionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiquidacionResponse'
        '404':
          description: '`LIQUIDACION_NO_ENCONTRADA`'
        '409':
          description: Conflicto de estado de la liquidación
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AnulacionRequest:
      properties:
        motivo:
          type: string
          maxLength: 500
          minLength: 5
          title: Motivo
      type: object
      required:
      - motivo
      title: AnulacionRequest
...
    CorreccionRequest:
      properties:
        fecha:
          type: string
          format: date
          title: Fecha
        ruta_id:
          type: string
          format: uuid
          title: Ruta Id
        vendedor_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Vendedor Id
        lineas:
          items:
            $ref: '#/components/schemas/LineaLiquidacion'
          type: array
          minItems: 1
          title: Lineas
        pagos:
          $ref: '#/components/schemas/PagosLiquidacion-Input'
        observaciones:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Observaciones
        motivo:
          type: string
          maxLength: 500
          minLength: 5
          title: Motivo
      type: object
      required:
      - fecha
      - ruta_id
      - lineas
      - motivo
      title: CorreccionRequest
...
    CuadreUnidades:
      properties:
        cargadas:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Cargadas
        vendidas:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Vendidas
        devueltas:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Devueltas
        merma:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Merma
        diferencia:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Diferencia
        cuadra:
          type: boolean
          title: Cuadra
      type: object
      required:
      - cargadas
      - vendidas
      - devueltas
      - merma
      - diferencia
      - cuadra
      title: CuadreUnidades
...
    DemandModelInfo:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        algoritmo:
          type: string
          title: Algoritmo
        version:
          type: string
          title: Version
      type: object
      required:
      - id
      - algoritmo
      - version
      title: DemandModelInfo
...
    DemandProductForecast:
      properties:
        producto_id:
          type: string
          format: uuid
          title: Producto Id
        serie:
          items:
            $ref: '#/components/schemas/DemandPoint'
          type: array
          title: Serie
      type: object
      required:
      - producto_id
      - serie
      title: DemandProductForecast
    DemandRequest:
      properties:
        producto_ids:
          items:
            type: string
            format: uuid
          type: array
          maxItems: 200
          minItems: 1
          title: Producto Ids
        ruta_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Ruta Id
          description: Si se omite, la demanda es agregada (suma de todas las rutas)
        fecha_base:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Fecha Base
          description: Por defecto hoy
        horizonte_dias:
          type: integer
          maximum: 30.0
          minimum: 1.0
          title: Horizonte Dias
        incluir_intervalo:
          type: boolean
          title: Incluir Intervalo
          default: true
        persistir:
          type: boolean
          title: Persistir
          description: Guarda los pronósticos en la BD
          default: false
      type: object
      required:
      - producto_ids
      - horizonte_dias
      title: DemandRequest
    DemandResponse:
      properties:
        modelo:
          $ref: '#/components/schemas/DemandModelInfo'
        generado_en:
          type: string
          format: date-time
          title: Generado En
        pronosticos:
          items:
            $ref: '#/components/schemas/DemandProductForecast'
          type: array
          title: Pronosticos
      type: object
      required:
      - modelo
      - generado_en
      - pronosticos
      title: DemandResponse
...
    EstadoLiquidacion:
      type: string
      enum:
      - borrador
      - cerrada
      - anulada
      title: EstadoLiquidacion
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LineaLiquidacion:
      properties:
        producto_id:
          type: string
          format: uuid
          title: Producto Id
        cantidad_cargada:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,10}|(?=[\d.]{1,13}0*$)\d{0,10}\.\d{0,2}0*$)
          title: Cantidad Cargada
        cantidad_vendida:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,10}|(?=[\d.]{1,13}0*$)\d{0,10}\.\d{0,2}0*$)
          title: Cantidad Vendida
        cantidad_devuelta:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,10}|(?=[\d.]{1,13}0*$)\d{0,10}\.\d{0,2}0*$)
          title: Cantidad Devuelta
          default: '0'
        cantidad_merma:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,10}|(?=[\d.]{1,13}0*$)\d{0,10}\.\d{0,2}0*$)
          title: Cantidad Merma
          default: '0'
        precio_unitario:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,12}|(?=[\d.]{1,15}0*$)\d{0,12}\.\d{0,2}0*$)
          - type: 'null'
          title: Precio Unitario
        agotado:
          type: boolean
          title: Agotado
          default: false
        justificacion_carga:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Justificacion Carga
      type: object
      required:
      - producto_id
      - cantidad_cargada
      - cantidad_vendida
      title: LineaLiquidacion
...
    LiquidacionConfig-Input:
      properties:
        umbral_diferencia_caja:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,12}|(?=[\d.]{1,15}0*$)\d{0,12}\.\d{0,2}0*$)
          title: Umbral Diferencia Caja
      type: object
      required:
      - umbral_diferencia_caja
      title: LiquidacionConfig
    LiquidacionConfig-Output:
      properties:
        umbral_diferencia_caja:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,12}|(?=[\d.]{1,15}0*$)\d{0,12}\.\d{0,2}0*$)
          title: Umbral Diferencia Caja
      type: object
      required:
      - umbral_diferencia_caja
      title: LiquidacionConfig
    LiquidacionRequest:
      properties:
        fecha:
          type: string
          format: date
          title: Fecha
        ruta_id:
          type: string
          format: uuid
          title: Ruta Id
        vendedor_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Vendedor Id
        lineas:
          items:
            $ref: '#/components/schemas/LineaLiquidacion'
          type: array
          minItems: 1
          title: Lineas
        pagos:
          $ref: '#/components/schemas/PagosLiquidacion-Input'
        observaciones:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Observaciones
      type: object
      required:
      - fecha
      - ruta_id
      - lineas
      title: LiquidacionRequest
    LiquidacionResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        fecha:
          type: string
          format: date
          title: Fecha
        ruta_id:
          type: string
          format: uuid
          title: Ruta Id
        ruta_nombre:
          type: string
          title: Ruta Nombre
        vendedor_id:
          type: string
          format: uuid
          title: Vendedor Id
        vendedor_nombre:
          type: string
          title: Vendedor Nombre
        estado:
          $ref: '#/components/schemas/EstadoLiquidacion'
        version:
          type: integer
          title: Version
        lote_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Lote Id
        lineas:
          items:
            $ref: '#/components/schemas/LineaRespuesta'
          type: array
          title: Lineas
        pagos:
          $ref: '#/components/schemas/PagosLiquidacion-Output'
        cuadre_unidades:
          $ref: '#/components/schemas/CuadreUnidades'
        cuadre_dinero:
          $ref: '#/components/schemas/CuadreDinero'
        devolucion_esperada:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Devolucion Esperada
        observaciones:
          anyOf:
          - type: string
          - type: 'null'
          title: Observaciones
        advertencias:
          items:
            type: string
          type: array
          title: Advertencias
        alerta_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Alerta Id
        creado_en:
          type: string
          format: date-time
          title: Creado En
        cerrado_en:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Cerrado En
        anulado_en:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Anulado En
        anulado_motivo:
          anyOf:
          - type: string
          - type: 'null'
          title: Anulado Motivo
      type: object
      required:
      - id
      - fecha
      - ruta_id
      - ruta_nombre
      - vendedor_id
      - vendedor_nombre
      - estado
      - version
      - lineas
      - pagos
      - cuadre_unidades
      - cuadre_dinero
      - devolucion_esperada
      - creado_en
      title: LiquidacionResponse
...
    ListadoLiquidaciones:
      properties:
        total:
          type: integer
          title: Total
        liquidaciones:
          items:
            $ref: '#/components/schemas/LiquidacionResumen'
          type: array
          title: Liquidaciones
      type: object
      required:
      - total
      - liquidaciones
      title: ListadoLiquidaciones
...
    PagosLiquidacion-Input:
      properties:
        efectivo:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,12}|(?=[\d.]{1,15}0*$)\d{0,12}\.\d{0,2}0*$)
          title: Efectivo
          default: '0'
        transferencia:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,12}|(?=[\d.]{1,15}0*$)\d{0,12}\.\d{0,2}0*$)
          title: Transferencia
          default: '0'
        credito:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,12}|(?=[\d.]{1,15}0*$)\d{0,12}\.\d{0,2}0*$)
          title: Credito
          default: '0'
        cobro_saldos:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,12}|(?=[\d.]{1,15}0*$)\d{0,12}\.\d{0,2}0*$)
          title: Cobro Saldos
          default: '0'
        gastos:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,12}|(?=[\d.]{1,15}0*$)\d{0,12}\.\d{0,2}0*$)
          title: Gastos
          default: '0'
        efectivo_entregado:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,12}|(?=[\d.]{1,15}0*$)\d{0,12}\.\d{0,2}0*$)
          title: Efectivo Entregado
          default: '0'
      type: object
      title: PagosLiquidacion
...
    PrecargaResponse:
      properties:
        fecha:
          type: string
          format: date
          title: Fecha
        ruta_id:
          type: string
          format: uuid
          title: Ruta Id
        vendedor_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Vendedor Id
        carga_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Carga Id
        lineas:
          items:
            $ref: '#/components/schemas/LineaPrecarga'
          type: array
          title: Lineas
        liquidacion:
          anyOf:
          - $ref: '#/components/schemas/LiquidacionResponse'
          - type: 'null'
        advertencias:
          items:
            type: string
          type: array
          title: Advertencias
      type: object
      required:
      - fecha
      - ruta_id
      - lineas
      title: PrecargaResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer